No need to have ninja.py which just references ninjaCommon, rename package to ninja and get rid of ninja.py
