+++ SCons/Tool/ninja/Globals.py
