+++ SCons/Tool/ninja/Rules.py
