+++ SCons/Tool/ninja/ninja.xml
